Move files around, implement localauth
