+++ rival_regions_wrapper/__init__.py
@@ -6,5 +6,5 @@ This unofficial API wrapper is an implementation
 of some Rival Regions functionalities.
 """
 
-# from .authentication_handler.authentication_handler import AuthenticationHandler
-from .middleware.middleware import LocalAuthentication, RemoteAuthentication
+from .authentication_handler import AuthenticationHandler
+from .middleware import LocalAuthentication, RemoteAuthentication
